+++ packages/keyring-controller/src/index.ts
@@ -26,6 +26,15 @@ export type {
   KeyringControllerWithKeyringV2Action,
   KeyringControllerWithKeyringV2UnsafeAction,
   KeyringControllerExportSeedPhraseAction,
+  KeyringControllerVerifyPasswordAction,
+  KeyringControllerExportAccountAction,
+  KeyringControllerImportAccountWithStrategyAction,
+  KeyringControllerSetLockedAction,
+  KeyringControllerChangePasswordAction,
+  KeyringControllerSubmitEncryptionKeyAction,
+  KeyringControllerExportEncryptionKeyAction,
+  KeyringControllerSubmitPasswordAction,
+  KeyringControllerGetAccountKeyringTypeAction,
 } from './KeyringController-method-action-types';
 export type * from './types';
 export * from './errors';