+++ packages/transaction-pay-controller/src/index.ts
@@ -1,6 +1,7 @@
 export type {
   TransactionConfig,
   TransactionConfigCallback,
+  TransactionData,
   TransactionFiatPayment,
   TransactionFiatPaymentCallback,
   TransactionPayControllerActions,